feat(connector): move copier to connector
